requirements.txt
transformers>=4.30.0
torch>=2.0.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langgraph>=0.0.10
python-dotenv>=1.0.0
langchain-community>=0.0.10